Start moving Ocr3 code
